move reads for BooleanExpr into one place
